Fixed midair jump and relative motion
Fixed an issue where the player could still jump if they walked off a platform instead of jumping, caused by isGrounded not being reset correctly. Also fixed an issue where the player would not move with a horizontal platform that they were standing on, bypassing actual physics and the understanding of all platformers. Caused by not actually being a thing in the first place.