Fixed enemy spawnrate ramping over time
Fixed an issue where enemies would spawn at an accelerated rate as time went on. Had nothing to do with time per se, it was actually an issue with the projectiles wrapping back around and never despawning. This meant that as time went on, more and more projectiles would spawn in without the old ones leaving, causing a constant increase in projectiles